copy: add slug_perm to output
When using the copy command to copy packages between repositories, the unique ID
(slug_perm) of the package wasn't being printed to stdout.  As with other commands
(such as 'push') which did, the output of this could then be parsed to further operate
on the specific package.

Therefore, make the output from 'copy' behave in the same way.